requirements.txt
# ------------------------------------------------------------
# Producer Python Dependencies (pinned)
# Purpose: deterministic build for the Kafka producer container.
# Python: 3.9
# ------------------------------------------------------------
kafka-python==2.2.15
Faker==37.6.0
